Refactored all the code from App.js to Home.js and wrapped the app code in navigation container
